Initial commit of Coding with Chrome BETA!
Thank you for trying our product, we are very excited to put it in front to you.

Features of this release are:
- IDE supporting Javascript, Coffeescript and Blockly
- Support for LEGO(R) MINDSTORMS(R) robots / EV3
- more to come soon ...

We are proud of what we have here, because it should be functional and solid but it needs your help!

First - we need feedback - what works, what doesn't, and most importantly: what do you think it needs for YOU to start using it?
Please post issues to https://github.com/google/coding-with-chrome/issues and we, or someone from the community will hopefully take it into account.

Second - if you want to contribute, check for details in the Contributing.md file

- The Coding with Chrome team.