Improved internal HTTP server and moved static files into Cache instance.
This change introduce the "cache" which is only filled / updated if the local cache version is lower then the current version.

The cache is used by the HTTP server and several other components.
The main benefit of this change is the lower payload for the sandbox (some urls instead of lots of data) and the faster loading of the application.
The cache is using IndexDB for storing the data which also lowers the memory  usage for storing these data.