Code cleanup and modern integration of different parts.
This change introduce same changes to the code base and get rid of some deprecated functions and modules.
Its using the "cache" instance for all framework files and the user configrations which allows and faster start time and less resource usage.
'The new way the "render" is working allows to make sure that all preparations are down before the code is executed.
The first run time (init) is increased by about 0-5% but the follow up run times are decreased by 20%-30%.
This change is not ready for producation yet, until the overall code coverage is over 50%.