Refactored mode configruation and menu bar device buttons.
This changes introduces an better UI experience and only display relevant options depending on the selected mode.
It's also introduce the possibility to add mode specific connect handler to the menu bar buttons to simplified the connection.
The additional buttons are automatically disabled if the feature is not used or on the select screen to avoid any confusions.