Added example code support to Workbench
The tutorial feature supports example code that loads into the editor
with each step. This adds that support to the Workbench add-on too by
treating a .cwc file attached to a Workbench project step as example
code.

Files are identified by their extension in the URL, but but the
underlying CwC file must also have at least one embedded content with
a mime type that matches the mime type of the mode loaded into the
editor, which is determined by the Workbench CwC tag attached to the
project.